Rename/Youtude컴포넌트 component로 이동
